+++ gitlab_tests/test_users.py
@@ -2,7 +2,7 @@ import responses
 from requests.exceptions import HTTPError
 
 from gitlab import Gitlab
-from gitlab_tests.base import BaseTest
+from gitlab_tests.base_test import BaseTest
 from response_data.users import *
